refactor

+++ src/features/blogs/api/blogs.controller.ts
@@ -4,6 +4,7 @@ import {
   Delete,
   Get,
   HttpCode,
+  HttpException,
   HttpStatus,
   Param,
   Post,
@@ -20,11 +21,13 @@ import { AddPostInputModel } from '../../posts/api/models/input/add-post.input.m
 import { PostsQueryRepo } from '../../posts/infrastructure/posts.query-repo';
 import { PostsQueryParamsInputModel } from '../../posts/api/models/input/posts-query-params.input.model';
 import { setPagination } from '../../../utils/set-pagination';
+import { BlogsRepo } from '../infrastructure/blogs.repo';
 
 @Controller('blogs')
 export class BlogsController {
   constructor(
     private blogsService: BlogsService,
+    private blogsRepo: BlogsRepo,
     private blogsQueryRepo: BlogsQueryRepo,
     private postsQueryRepo: PostsQueryRepo
   ) {}
@@ -39,6 +42,11 @@ export class BlogsController {
   @Get(':id')
   @HttpCode(HttpStatus.OK)
   async getBlogById(@Param('id') id: string) {
+    const existsBlog = await this.blogsRepo.existsById(id);
+
+    if (!existsBlog) {
+      throw new HttpException(`Blog with id ${id} not found`, HttpStatus.NOT_FOUND);
+    }
     return await this.blogsQueryRepo.findById(id);
   }
   @Get(':id/posts')
@@ -48,6 +56,11 @@ export class BlogsController {
     @Query()
     queryParams: PostsQueryParamsInputModel
   ) {
+    const existsBlog = await this.blogsRepo.existsById(id);
+
+    if (!existsBlog) {
+      throw new HttpException(`Blog with id ${id} not found`, HttpStatus.NOT_FOUND);
+    }
     return await this.postsQueryRepo.findAll(setPagination(queryParams), id);
   }
   @Post()
@@ -60,18 +73,33 @@ export class BlogsController {
   @Post(':id/posts')
   @HttpCode(HttpStatus.CREATED)
   async addPostSpecificBlog(@Param('id') id: string, @Body() addPostInputModel: AddPostInputModel) {
+    const existsBlog = await this.blogsRepo.existsById(id);
+
+    if (!existsBlog) {
+      throw new HttpException(`Blog with id ${id} not found`, HttpStatus.NOT_FOUND);
+    }
     const addedPostId = await this.blogsService.addPostSpecificBlog(id, addPostInputModel);
     return this.postsQueryRepo.findById(addedPostId);
   }
   @Put(':id')
   @HttpCode(HttpStatus.NO_CONTENT)
   async updateBlog(@Param('id') id: string, @Body() updateBlogModel: UpdateBlogInputModel) {
+    const existsBlog = await this.blogsRepo.existsById(id);
+
+    if (!existsBlog) {
+      throw new HttpException(`Blog with id ${id} not found`, HttpStatus.NOT_FOUND);
+    }
     await this.blogsService.updateBlog(id, updateBlogModel);
   }
 
   @Delete(':id')
   @HttpCode(HttpStatus.NO_CONTENT)
   async deleteBlog(@Param('id') id: string) {
+    const existsBlog = await this.blogsRepo.existsById(id);
+
+    if (!existsBlog) {
+      throw new HttpException(`Blog with id ${id} not found`, HttpStatus.NOT_FOUND);
+    }
     await this.blogsService.deleteBlog(id);
   }
 }
